[neptune/#124] Improve Invite User functionality Refactor: Make User\EditJurisdictions into a shared view so it can be used to meet the new invite user requirements
