+++ Source/Neptune.Web/Neptune.Web.csproj
@@ -227,6 +227,7 @@
     <Compile Include="Models\Generated\WaterQualityManagementPlanPriorityPrimaryKey.Binding.cs" />
     <Compile Include="Models\Generated\WaterQualityManagementPlanStatus.Binding.cs" />
     <Compile Include="Models\Generated\WaterQualityManagementPlanStatusPrimaryKey.Binding.cs" />
+    <Compile Include="Views\Shared\UserJurisdictions\EditUserJurisdictions.cs" />
     <Compile Include="Views\WaterQualityManagementPlan\EditWqmpTreatmentBmps.cs" />
     <Compile Include="Security\WaterQualityManagementPlanCreateFeature.cs" />
     <Compile Include="Security\WaterQualityManagementPlanDocumentManageFeature.cs" />
@@ -949,9 +950,8 @@
     <Compile Include="Views\TreatmentBMP\New.cs" />
     <Compile Include="Views\TreatmentBMP\NewViewData.cs" />
     <Compile Include="Views\TreatmentBMP\NewViewModel.cs" />
-    <Compile Include="Views\User\EditJurisdictions.cs" />
-    <Compile Include="Views\User\EditJurisdictionsViewData.cs" />
-    <Compile Include="Views\User\EditJurisdictionsViewModel.cs" />
+    <Compile Include="Views\Shared\UserJurisdictions\EditUserJurisdictionsViewData.cs" />
+    <Compile Include="Views\Shared\UserJurisdictions\EditUserJurisdictionsViewModel.cs" />
     <Compile Include="Views\Tenant\Detail.cs" />
     <Compile Include="Views\Home\Index.cs" />
     <Compile Include="Views\Home\IndexViewData.cs" />
@@ -1076,7 +1076,7 @@
     <Content Include="Views\FundingEvent\EditFundingEventFundingSourceController.js" />
     <Content Include="Views\TreatmentBMPType\EditTreatmentBMPTypeController.js" />
     <Content Include="Views\TreatmentBMP\TreatmentBMPMapController.js" />
-    <Content Include="Views\User\EditJurisdictionsController.js" />
+    <Content Include="Views\Shared\UserJurisdictions\EditUserJurisdictionsController.js" />
     <Content Include="Views\TreatmentBMPAssessment\ScoreTableController.js" />
     <Content Include="Views\Tenant\Detail.cshtml" />
     <Content Include="Views\Tenant\EditBasics.cshtml" />
@@ -1106,7 +1106,7 @@
     <Content Include="Views\ModeledCatchment\UpdateModeledCatchmentGeometry.cshtml" />
     <Content Include="Views\ModeledCatchment\UpdateModeledCatchmentGeometryErrors.cshtml" />
     <Content Include="Views\Assessment\Index.cshtml" />
-    <Content Include="Views\User\EditJurisdictions.cshtml" />
+    <Content Include="Views\Shared\UserJurisdictions\EditUserJurisdictions.cshtml" />
     <Content Include="Views\TreatmentBMPAssessmentObservationType\Manage.cshtml" />
     <Content Include="Views\TreatmentBMPAssessmentObservationType\Detail.cshtml" />
     <Content Include="Views\TreatmentBMPAssessmentObservationType\Edit.cshtml" />
