+++ Source/Neptune.Web/Views/Shared/UserJurisdictions/EditUserJurisdictions.cs
@@ -18,9 +18,9 @@ GNU Affero General Public License <http://www.gnu.org/licenses/> for more detail
 Source code is available upon request via <support@sitkatech.com>.
 </license>
 -----------------------------------------------------------------------*/
-namespace Neptune.Web.Views.User
+namespace Neptune.Web.Views.Shared.UserJurisdictions
 {
-    public abstract class EditJurisdictions : LtInfo.Common.Mvc.TypedWebPartialViewPage<EditJurisdictionsViewData, EditJurisdictionsViewModel>
+    public abstract class EditUserJurisdictions : LtInfo.Common.Mvc.TypedWebPartialViewPage<EditUserJurisdictionsViewData, EditUserJurisdictionsViewModel>
     {
     }
 }
