+++ Source/Neptune.Web/Views/Shared/UserJurisdictions/EditUserJurisdictionsController.js
@@ -18,7 +18,7 @@ GNU Affero General Public License <http://www.gnu.org/licenses/> for more detail
 Source code is available upon request via <support@sitkatech.com>.
 </license>
 -----------------------------------------------------------------------*/
-angular.module("NeptuneApp").controller("EditJurisdictionsController", function ($scope, angularModelAndViewData) {
+angular.module("NeptuneApp").controller("EditUserJurisdictionsController", function ($scope, angularModelAndViewData) {
     $scope.AngularModel = angularModelAndViewData.AngularModel;
     $scope.AngularViewData = angularModelAndViewData.AngularViewData;
