+++ Source/Neptune.Web/Views/User/Detail.cshtml
@@ -30,7 +30,7 @@
 {
     @{ DhtmlxGridIncludes.RenderPartialView(Html); }
     <script src="@Url.Content("~/Views/PersonOrganization/PersonOrganizationController.js")" type="text/javascript"></script>
-    <script src="@Url.Content("~/Views/User/EditJurisdictionsController.js")" type="text/javascript"></script>
+    <script src="@Url.Content("~/Views/Shared/UserJurisdictions/EditUserJurisdictionsController.js")" type="text/javascript"></script>
 
     <style type="text/css">
         .paddedTableCell {
